Make re-enabling button after `stop()` optional
Adds an optional boolean parameter to the `stop()` function, which if
`true` does not run the jQuery to remove the disabled attribute from the
button (i.e. the button remains disabled) after calling `stop()`. This
allows the enabled/disabled state of the button to be determined by a
reactive datasource, e.g. a template helper.

Call `stop(true)` on return from Meteor method:

    Meteor.call('myMethod', function() {
      Meteor.setTimeout(function() {
        laddaButton.stop(true);
      }, 100);
    });

Helper, `laddaEnabled()` that determines enabled/disabled state:

    Template.myTemplate.helpers({
      laddaEnabled: function() {
        return (/* some logic */) ? true : false;
      }
    });

Usage in template:

    <button class="ladda-button" disabled={{laddaEnabled}} data-style="expand-right"><span class="ladda-label">Submit</span></button>